# Binary encoding for externs