--- a/lib/purescript-ast/package.yaml
+++ b/lib/purescript-ast/package.yaml
@@ -29,6 +29,7 @@ dependencies:
   - mtl >=2.1.0 && <2.3.0
   - protolude >=0.1.6 && <0.2.4
   - scientific >=0.3.4.9 && <0.4
+  - store
   - text
   - vector