--- a/lib/purescript-ast/src/Language/PureScript/AST/Declarations.hs
+++ b/lib/purescript-ast/src/Language/PureScript/AST/Declarations.hs
@@ -8,6 +8,7 @@ module Language.PureScript.AST.Declarations where
 
 import Prelude.Compat
 
+import Codec.Serialise (Serialise)
 import Control.DeepSeq (NFData)
 import Data.Functor.Identity
 
@@ -180,7 +181,7 @@ data DeclarationRef
   -- elaboration in name desugaring.
   --
   | ReExportRef SourceSpan ExportSource DeclarationRef
-  deriving (Show, Generic, NFData, Store)
+  deriving (Show, Generic, NFData, Store, Serialise)
 
 instance Eq DeclarationRef where
   (TypeRef _ name dctors) == (TypeRef _ name' dctors') = name == name' && dctors == dctors'
@@ -198,7 +199,7 @@ data ExportSource =
   { exportSourceImportedFrom :: Maybe ModuleName
   , exportSourceDefinedIn :: ModuleName
   }
-  deriving (Eq, Ord, Show, Generic, NFData, Store)
+  deriving (Eq, Ord, Show, Generic, NFData, Store, Serialise)
 
 -- enable sorting lists of explicitly imported refs when suggesting imports in linting, IDE, etc.
 -- not an Ord because this implementation is not consistent with its Eq instance.
@@ -283,7 +284,7 @@ data ImportDeclarationType
   -- An import with a list of references to hide: `import M hiding (foo)`
   --
   | Hiding [DeclarationRef]
-  deriving (Eq, Show, Generic, Store)
+  deriving (Eq, Show, Generic, Store, Serialise)
 
 isImplicit :: ImportDeclarationType -> Bool
 isImplicit Implicit = True