--- a/lib/purescript-ast/src/Language/PureScript/AST/Operators.hs
+++ b/lib/purescript-ast/src/Language/PureScript/AST/Operators.hs
@@ -9,6 +9,7 @@ import GHC.Generics (Generic)
 import Control.DeepSeq (NFData)
 import Data.Aeson ((.=))
 import qualified Data.Aeson as A
+import Data.Store (Store)
 
 import Language.PureScript.Crash
 
@@ -24,6 +25,7 @@ data Associativity = Infixl | Infixr | Infix
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData Associativity
+instance Store Associativity
 
 showAssoc :: Associativity -> String
 showAssoc Infixl = "infixl"
@@ -49,6 +51,7 @@ data Fixity = Fixity Associativity Precedence
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData Fixity
+instance Store Fixity
 
 instance A.ToJSON Fixity where
   toJSON (Fixity associativity precedence) =