--- a/lib/purescript-ast/src/Language/PureScript/AST/SourcePos.hs
+++ b/lib/purescript-ast/src/Language/PureScript/AST/SourcePos.hs
@@ -6,6 +6,7 @@ module Language.PureScript.AST.SourcePos where
 
 import Prelude.Compat
 
+import Codec.Serialise (Serialise)
 import Control.DeepSeq (NFData)
 import Data.Aeson ((.=), (.:))
 import Data.Text (Text)
@@ -25,7 +26,7 @@ data SourcePos = SourcePos
     -- ^ Line number
   , sourcePosColumn :: Int
     -- ^ Column number
-  } deriving (Show, Eq, Ord, Generic, NFData, Store)
+  } deriving (Show, Eq, Ord, Generic, NFData, Store, Serialise)
 
 displaySourcePos :: SourcePos -> Text
 displaySourcePos sp =
@@ -53,7 +54,7 @@ data SourceSpan = SourceSpan
     -- ^ Start of the span
   , spanEnd :: SourcePos
     -- ^ End of the span
-  } deriving (Show, Eq, Ord, Generic, NFData, Store)
+  } deriving (Show, Eq, Ord, Generic, NFData, Store, Serialise)
 
 displayStartEndPos :: SourceSpan -> Text
 displayStartEndPos sp =