--- a/lib/purescript-ast/src/Language/PureScript/Comments.hs
+++ b/lib/purescript-ast/src/Language/PureScript/Comments.hs
@@ -10,7 +10,6 @@ import Codec.Serialise (Serialise)
 import Control.DeepSeq (NFData)
 import Data.Text (Text)
 import GHC.Generics (Generic)
-import Data.Store (Store)
 
 import Data.Aeson.TH
 
@@ -20,7 +19,6 @@ data Comment
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData Comment
-instance Store Comment
 instance Serialise Comment
 
 $(deriveJSON (defaultOptions { sumEncoding = ObjectWithSingleField }) ''Comment)