--- a/lib/purescript-ast/src/Language/PureScript/Names.hs
+++ b/lib/purescript-ast/src/Language/PureScript/Names.hs
@@ -14,6 +14,7 @@ import Data.Functor.Contravariant (contramap)
 import GHC.Generics (Generic)
 import Data.Aeson
 import Data.Aeson.TH
+import Data.Store (Store)
 import Data.Text (Text)
 import qualified Data.Text as T
 
@@ -29,6 +30,7 @@ data Name
   deriving (Eq, Ord, Show, Generic)
 
 instance NFData Name
+instance Store Name
 
 getIdentName :: Name -> Maybe Ident
 getIdentName (IdentName name) = Just name
@@ -77,6 +79,7 @@ data Ident
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData Ident
+instance Store Ident
 
 runIdent :: Ident -> Text
 runIdent (Ident i) = i
@@ -100,6 +103,7 @@ newtype OpName (a :: OpNameType) = OpName { runOpName :: Text }
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData (OpName a)
+instance Store (OpName a)
 
 instance ToJSON (OpName a) where
   toJSON = toJSON . runOpName
@@ -125,6 +129,7 @@ newtype ProperName (a :: ProperNameType) = ProperName { runProperName :: Text }
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData (ProperName a)
+instance Store (ProperName a)
 
 instance ToJSON (ProperName a) where
   toJSON = toJSON . runProperName
@@ -156,6 +161,7 @@ newtype ModuleName = ModuleName [ProperName 'Namespace]
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData ModuleName
+instance Store ModuleName
 
 runModuleName :: ModuleName -> Text
 runModuleName (ModuleName pns) = T.intercalate "." (runProperName <$> pns)
@@ -179,6 +185,7 @@ data Qualified a = Qualified (Maybe ModuleName) a
   deriving (Show, Eq, Ord, Functor, Foldable, Traversable, Generic)
 
 instance NFData a => NFData (Qualified a)
+instance Store a => Store (Qualified a)
 
 showQualified :: (a -> Text) -> Qualified a -> Text
 showQualified f (Qualified Nothing a) = f a