--- a/lib/purescript-ast/src/Language/PureScript/PSString.hs
+++ b/lib/purescript-ast/src/Language/PureScript/PSString.hs
@@ -11,6 +11,7 @@ module Language.PureScript.PSString
 
 import Prelude.Compat
 import GHC.Generics (Generic)
+import Codec.Serialise (Serialise)
 import Control.DeepSeq (NFData)
 import Control.Exception (try, evaluate)
 import Control.Applicative ((<|>))
@@ -52,6 +53,7 @@ newtype PSString = PSString { toUTF16CodeUnits :: [Word16] }
 
 instance NFData PSString
 instance Store PSString
+instance Serialise PSString
 
 instance Show PSString where
   show = show . codePoints