--- a/lib/purescript-ast/src/Language/PureScript/Roles.hs
+++ b/lib/purescript-ast/src/Language/PureScript/Roles.hs
@@ -13,6 +13,7 @@ import Control.DeepSeq (NFData)
 import qualified Data.Aeson as A
 import qualified Data.Aeson.TH as A
 import GHC.Generics (Generic)
+import Data.Store (Store)
 
 -- |
 -- The role of a type constructor's parameter.
@@ -29,5 +30,6 @@ data Role
   deriving (Show, Eq, Ord, Generic)
 
 instance NFData Role
+instance Store Role
 
 $(A.deriveJSON A.defaultOptions ''Role)