--- a/lib/purescript-ast/src/Language/PureScript/Roles.hs
+++ b/lib/purescript-ast/src/Language/PureScript/Roles.hs
@@ -9,6 +9,7 @@ module Language.PureScript.Roles
 
 import Prelude.Compat
 
+import Codec.Serialise (Serialise)
 import Control.DeepSeq (NFData)
 import qualified Data.Aeson as A
 import qualified Data.Aeson.TH as A
@@ -31,5 +32,6 @@ data Role
 
 instance NFData Role
 instance Store Role
+instance Serialise Role
 
 $(A.deriveJSON A.defaultOptions ''Role)