… package.yaml
… psc-ide/README.md
… src/Language/PureScript/Externs.hs
… src/Language/PureScript/Ide.hs
… src/Language/PureScript/Ide/Externs.hs
… src/Language/PureScript/Make.hs
… src/Language/PureScript/Make/Actions.hs
… src/Language/PureScript/Make/Monad.hs
… stack.yaml