I did some profiling on the ide load times, and everything was dominated by JSON parsing. So I tried deriving `Store` (a binary serialize lib) for all the data types necessary for ExternsFile and then using the binary encoding I get significant speedups. Here are the stats for `pscid`:

Before:
```
	   ide-bench.exe +RTS -N -p -RTS
	total time  =        3.49 secs   (12947 ticks @ 1000 us, 8 processors)
	total alloc = 8,460,187,856 bytes  (excludes profiling overheads)
```
After:
```
	   ide-bench.exe +RTS -N -p -RTS
	total time  =        0.50 secs   (1837 ticks @ 1000 us, 8 processors)
	total alloc = 2,664,745,016 bytes  (excludes profiling overheads)
```

I'm just opening this as a draft for now. Do we want to make the switch to a binary encoding? Do we want to use a different library? I've found `binary`, `cereal` and `store` so far and picked `store` because it had both TH as well as Generic deriving. It's possible we could see some speedups in incremental builds as well, but any sizeable project I have generates too many warnings from the changes in master to give good timings right now.