@garyb 2020-04-10
I have no problem with externs being binary personally, there's no real advantage to those files being human readable that I can think of.