@hdgarrood 2020-04-10
Yeah, I always thought we would want to switch to binary at some point too.