@hdgarrood 2020-04-10
I'm not terribly concerned about the choice of library; since the externs file format is internal we can always change it later if we find a different library is better suited.