@natefaubion 2020-04-10
Related, I've created #3842 to track the bloat issues with externs.