@joneshf 2020-04-11
Might also want to look into [`cborg`](https://hackage.haskell.org/package/cborg). In some informal testing it ended up being about twice as fast as `binary` or `store`. Not sure if that was a fluke or I was using the others incorrectly. But it's a data point.