@kritzcreek 2020-04-11
New findings!
> Might also want to look into cborg

Thanks for that recommendation @joneshf. The `cborg` library and the stuff built on top of it are documented much better and I'd prefer to use an established binary format like `cbor` so if someone wants to hack together some tooling in, let's say, Rust they can do so.

The CBOR standard actually defines a bijection to JSON and https://hackage.haskell.org/package/cborg-json-0.2.2.0 implements it. So one thing we could do is continue using all our existing JSON instances (this also addresses @hdgarrood's concern about the version field) and put a little JSON <-> CBOR adapter in between reading and writing from the file system. Performance-wise this puts us somewhere in the middle between JSON and Store. It's 3 times faster than just using JSON, and two times slower than going for direct Store instances.

```
	   ide-bench.exe +RTS -N -p -RTS

	total time  =        0.96 secs   (3547 ticks @ 1000 us, 8 processors)
	total alloc = 5,209,416,280 bytes  (excludes profiling overheads)
```

This approach still has all the problems outlined by @natefaubion though, we're encoding the same filepath over and over, and even worse when we're reading the file back in we get no sharing.

Next I'll try to use https://hackage.haskell.org/package/serialise (typeclass based encoders/decoders on top of cborg) to get a fair comparison between `cborg` and `store` and if they're competitive I'll give interning the file paths a try.