@kritzcreek 2020-04-11
Sad update :(

I derived instances for `Serialise`, but they don't seem to roundtrip. Whenever I try to decode a previously encoded `.cbor` file I get `DeserialiseFailure 1 "unknown tag"` as the only error output, which also gives me no way of figuring out where things went wrong.