@kritzcreek 2020-04-12
New Update.... type-class based encoding + decoding failures as Exceptions tricked me (It tried to deserialise values of type `Maybe ExternsFile`). With the Serialise codecs I'm now getting comparable timings to the `Store` en/decoding.

```
	   ide-bench.exe +RTS -N -p -RTS

	total time  =        0.55 secs   (2020 ticks @ 1000 us, 8 processors)
	total alloc = 1,649,605,704 bytes  (excludes profiling overheads)
```

Since CBOR is a self-describing format I'll see if I can extract the version from it without trying to decode the whole ExternsFile.