@kritzcreek 2020-04-12
Allright, I think this is in a reviewable state now.

In this PR I've changed the encoding for externs from JSON to CBOR. I'm using the `serialise` library to derive instances for all the relevant data types.

By doing this we're getting significant speedups in IDE startup time, but also in the normal compiler pipeline.
The performance measurements in this PR measure how long it takes to load the fully built `pscid` project and populate the IDE caches. It's heavily dominated by the time it takes to deserialise the externs files, and drops from 3.5s to 0.55s. Memory allocations drop from `8,460,187,856 bytes` to `1,649,605,704 bytes`.
There are more low-hanging fruits to further speed this up (some of which have already been identified in this PR), but I think this is a good starting point.

Because we're serialising to CBOR, it's possible to extract information like the version number that was stored with the externs, without having to be able to decode the whole format. Not using JSON anymore doesn't remove any features we used to have, because of this property. We will break downstream users that relied on the JSON format, but as far as I can tell we've agreed that externs are a compiler internal format for which we are not giving stability guarantees in any way.