@kritzcreek 2020-04-12
Thanks for the review @hdgarrood! I'll merge this and as mentioned earlier we can do more improvements in follow-up PRs.